[Sprite] quickfix for z ordering problems
+ in spriter objects can change z order, in Unity it’s impossible to
animate that
+ as a temporary solution, each object is distinguished by it’s
rendering group attached to it’s name. it makes him unique and is
reused ONLY if rendering group is also the same
+ this solves problem with die animation of the default test character
from Spriter essential pack

! this generates much more Unity nodes than it should, a better
solution would be taking all of these z orders and assigning new ones,
that are ok for all animations at once. also then, when some objects
are changing order in some animations, node can be duplicated just for
it

! this is a temporary fix, whenever i get some more time i’ll fix it
properly